split st.memo caching logic into memo_caches; add "underscore-prefixed argument" hash ignoring
