+++ lib/tests/streamlit/memo/memo_test.py
@@ -128,3 +128,39 @@ class MemoTest(unittest.TestCase):
 
         # The other thread should not have modified the main thread
         self.assertEqual(1, get_counter())
+
+    def test_ignored_args(self):
+        """Args prefixed with _ are not used as part of the cache key."""
+        call_count = [0]
+
+        @st.memo
+        def foo(arg1, _arg2, *args, kwarg1, _kwarg2=None, **kwargs):
+            call_count[0] += 1
+
+        foo(1, 2, 3, kwarg1=4, _kwarg2=5, kwarg3=6, _kwarg4=7)
+        self.assertEqual([1], call_count)
+
+        # Call foo again, but change the values for _arg2, _kwarg2, and _kwarg4.
+        # The call count shouldn't change, because these args will not be part
+        # of the hash.
+        foo(1, None, 3, kwarg1=4, _kwarg2=None, kwarg3=6, _kwarg4=None)
+        self.assertEqual([1], call_count)
+
+        # Changing the value of any other argument will increase the call
+        # count. We test each argument type:
+
+        # arg1 (POSITIONAL_OR_KEYWORD)
+        foo(None, 2, 3, kwarg1=4, _kwarg2=5, kwarg3=6, _kwarg4=7)
+        self.assertEqual([2], call_count)
+
+        # *arg (VAR_POSITIONAL)
+        foo(1, 2, None, kwarg1=4, _kwarg2=5, kwarg3=6, _kwarg4=7)
+        self.assertEqual([3], call_count)
+
+        # kwarg1 (KEYWORD_ONLY)
+        foo(1, 2, 3, kwarg1=None, _kwarg2=5, kwarg3=6, _kwarg4=7)
+        self.assertEqual([4], call_count)
+
+        # **kwarg (VAR_KEYWORD)
+        foo(1, 2, 3, kwarg1=4, _kwarg2=5, kwarg3=None, _kwarg4=7)
+        self.assertEqual([5], call_count)
