+++ memo_tests/memo_test.py
@@ -5,8 +5,8 @@ from typing import List, Optional, Callable, Any
 import streamlit as st
 
 
-def foo(a: int, b: int, *args, kw_only, **kwargs) -> str:
-    return f"{a}, {b}, {args}, {kw_only}, {kwargs}"
+def foo(posOnly, /, pos0, pos1, *args, kw_only, **kwargs) -> str:
+    return f"{posOnly}, {pos0}, {pos1}, {args}, {kw_only}, {kwargs}"
 
 
 signature = inspect.signature(foo)
@@ -14,7 +14,7 @@ params: List[inspect.Parameter] = list(signature.parameters.values())
 st.write(signature.parameters.values())
 
 
-def _get_arg_name(func: Callable[..., Any], index: int) -> Optional[str]:
+def _get_positional_arg_name(func: Callable[..., Any], index: int) -> Optional[str]:
     if index < 0:
         return None
 
@@ -22,7 +22,10 @@ def _get_arg_name(func: Callable[..., Any], index: int) -> Optional[str]:
     if index >= len(params):
         return None
 
-    if params[index].kind == inspect.Parameter.POSITIONAL_OR_KEYWORD:
+    if params[index].kind in (
+        inspect.Parameter.POSITIONAL_OR_KEYWORD,
+        inspect.Parameter.POSITIONAL_ONLY,
+    ):
         return params[index].name
 
     return None
@@ -33,17 +36,28 @@ def wrapper(func):
     def wrapped_func(*args, **kwargs):
         arg_pairs = []
         for arg_idx in range(len(args)):
-            arg_pairs.append((_get_arg_name(func, index=arg_idx), args[arg_idx]))
+            arg_pairs.append(
+                (_get_positional_arg_name(func, index=arg_idx), args[arg_idx])
+            )
         for kw_name, kw_val in kwargs.items():
             arg_pairs.append((kw_name, kw_val))
 
         arg_names = ", ".join(str(ap) for ap in arg_pairs)
-        st.write(f"{func.__name__} called with {arg_names}")
+        st.write(f"{func.__name__} called with `{arg_names}`")
 
         return func(*args, **kwargs)
 
     return wrapped_func
 
 
-result = wrapper(foo)(1, 2, 3, 4, kw_only=True, qwert=False)
+result = wrapper(foo)(
+    "posOnly",
+    "pos0",
+    "pos1",
+    "*args0",
+    "*args1",
+    kw_only="kw_only",
+    kwargs0="kwargs0",
+    kwargs1="kwargs1",
+)
 st.write(result)
